+++ src/Exception/AuthenticationFailed.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of Chrome PHP.
+ *
+ * (c) Soufiane Ghzal <sghzal@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace HeadlessChromium\Exception;
+
+use Exception;
+
+class AuthenticationFailed extends Exception
+{
+}