+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `disableJavascript` option
 * Allow choosing the box model in `Node::getPosition()`
 * Allow setting `referrer`, `referrerPolicy` and `transitionType` in `Page::navigate()`
+* Add `Page::authenticate` and `Page::clearAuthentication` methods
 
 
 ## 1.15.1 (UPCOMING)